backend/requirements.txt
# Core Framework
fastapi==0.109.0
uvicorn==0.27.0
python-multipart==0.0.6

# AI & Embeddings
anthropic==0.18.1
chromadb==0.4.22
sentence-transformers==2.2.2

# Document Processing
PyPDF2==3.0.1
python-docx==1.1.0
python-pptx==0.6.23
markdown==3.5.2

# Database
sqlalchemy==2.0.25
aiosqlite==0.19.0

# Utilities
python-dotenv==1.0.0
pydantic==2.5.0
structlog==24.1.0
python-dateutil==2.8.2
pytz==2024.1

# HTTP Client
httpx==0.26.0
aiofiles==23.2.1
